backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
faiss-cpu
PyPDF2
sentence-transformers
pydantic
python-dotenv
